fix(form): change specifyTokenId to uint64_t and revert abilityName backfill

AI[0%] Human Fixed[0%] Human[100%] AI Adopted[0%]

+++ services/include/ams_mgr/form_ams_helper.h
@@ -151,14 +151,14 @@ public:
      *        execute-done callback.
      * @param param The insight intent execute param.
      * @param wantParams The want params carrying intentName and intentParams.
-     * @param specifyTokenId The token id used by ams permission checking, 0 means not specified.
-     * @param callerToken The caller ability token of the form host, same as router.
+     * @param specifyTokenId The specified token id for permission checking.
+     * @param callerToken The caller ability token.
      * @return Returns ERR_OK on success, others on failure.
      */
     ErrCode ExecuteIntentWithSpecifyTokenId(uint64_t key,
         const sptr<AbilityRuntime::InsightIntentHostClient> &insightIntentHostClient,
         const InsightIntentExecuteParam &param, const AAFwk::WantParams &wantParams,
-        uint32_t specifyTokenId, const sptr<IRemoteObject> &callerToken);
+        uint64_t specifyTokenId, const sptr<IRemoteObject> &callerToken);
 private:
     /**
      * @brief acquire a form ability manager, if it not existed,
