+++ test/mock/include/mock_ability_manager.h
@@ -327,7 +327,7 @@ public:
     int32_t ExecuteIntentWithSpecifyTokenId(uint64_t key,
         const sptr<IRemoteObject> &hostClient,
         const InsightIntentExecuteParam &param, const AAFwk::WantParams &wantParams,
-        uint32_t specifyTokenId = 0, const sptr<IRemoteObject> &callerToken = nullptr)
+        uint64_t specifyTokenId = 0, const sptr<IRemoteObject> &callerToken = nullptr)
     {
         return 0;
     }
@@ -833,7 +833,7 @@ public:
     int32_t ExecuteIntentWithSpecifyTokenId(uint64_t key,
         const sptr<IRemoteObject> &hostClient,
         const InsightIntentExecuteParam &param, const AAFwk::WantParams &wantParams,
-        uint32_t specifyTokenId = 0, const sptr<IRemoteObject> &callerToken = nullptr)
+        uint64_t specifyTokenId = 0, const sptr<IRemoteObject> &callerToken = nullptr)
     {
         return 0;
     }
